backend/requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
numpy
scikit-learn
pypdf
PyMuPDF
groq
huggingface_hub
llama-cpp-python
fastapi
uvicorn
python-multipart
faster-whisper
edge-tts
streamlit-mic-recorder
textblob
matplotlib
seaborn
torch
requests
aiohttp
asyncpg
redis
networkx
